Remove a wrong type argument in IO.interpret
In a compiler patch that changes the way type variables are maximised in
typedUnapply the code would throw the following type error:

    .handleError(t => Right((t.asLeft[IO[B]], limit - 1)))
    //                       ^^^^^^^^^^^^^^^
    // Found:    Either[Throwable, cats.effect.IO[B]]
    // Required: B
    //
    // where:    B is a type in methodinterpret with bounds >: Either[Throwable, A$1]

Turns out the right type argument for the RHS of Either is A, not IO[B],
and it's only because that type variable was instantiated to Any or
Nothing was it fine to use `IO[B]` instead.  Additionally there's some
kind of type inference interaction with the choice of Right.apply and
asRight, which I didn't look into further - I just switched to using asRight.